Use mixins for DBApiHook
