+++ providers/common/ai/src/airflow/providers/common/ai/utils/datasource.py
