+++ providers/common/ai/tests/system/common/ai/example_llm_sql_query_operator.py
@@ -18,8 +18,8 @@
 from __future__ import annotations
 
 from airflow import DAG
-from airflow.providers.common.ai.configs.datasource import DataSourceConfig
 from airflow.providers.common.ai.operators.llm_sql import LLMSQLQueryOperator
+from airflow.providers.common.ai.utils.datasource import DataSourceConfig
 from airflow.utils import timezone
 
 datasource_config_with_schema = DataSourceConfig(
