+++ providers/common/ai/tests/unit/common/ai/utils/test_mixin.py
@@ -18,22 +18,29 @@ from __future__ import annotations
 
 import pytest
 
-from airflow.providers.common.ai.configs.datasource import DataSourceConfig
+from airflow.providers.common.ai.utils.mixins import CommonAIHookMixin
+from airflow.sdk import Connection
 
 
-class TestDataSourceConfig:
-    def test_valid_config(self):
-        config = DataSourceConfig(
-            conn_id="fake_conn",
-            uri="postgres://",
-            table_name="test_table",
-            schema={"id": "int", "name": "string"},
+class CommonAIHookTestMixin(CommonAIHookMixin):
+    def __init__(self):
+        pass
+
+
+class TestCommonAIHookMixin:
+    @pytest.fixture(autouse=True)
+    def setup_connections(self, create_connection_without_db):
+
+        conn_postgres = Connection(
+            conn_id="postgres_default",
+            conn_type="postgres",
+            password="postgres_password",
+            host="postgres_host",
         )
-        assert config.conn_id == "fake_conn"
-        assert config.uri == "postgres://"
-        assert config.table_name == "test_table"
-        assert config.schema == {"id": "int", "name": "string"}
+        create_connection_without_db(conn_postgres)
 
-    def test_invalid_schema_type(self):
-        with pytest.raises(ValueError, match="Schema must be a dictionary of column names and types"):
-            DataSourceConfig(conn_id="test_conn", uri="test_uri", schema=["invalid_schema"])
+    def test_get_db_api_hook(self):
+        """Test to validate it fetches DBAPi based hooks"""
+        common_ai_hook_mixin = CommonAIHookTestMixin()
+        result = common_ai_hook_mixin.get_db_api_hook("postgres_default")
+        assert result.dialect_name == "postgresql"
