+++ pyproject.toml
@@ -105,9 +105,6 @@ packages = []
 "amazon" = [
     "apache-airflow-providers-amazon>=9.0.0"
 ]
-"apache.beam" = [
-    "apache-airflow-providers-apache-beam>=5.8.1; python_version !=\"3.13\""
-]
 "apache.cassandra" = [
     "apache-airflow-providers-apache-cassandra>=3.7.0"
 ]
@@ -399,7 +396,6 @@ packages = []
     "apache-airflow-providers-airbyte>=5.0.0",
     "apache-airflow-providers-alibaba>=3.0.0",
     "apache-airflow-providers-amazon>=9.0.0",
-    "apache-airflow-providers-apache-beam>=5.8.1; python_version !=\"3.13\"",
     "apache-airflow-providers-apache-cassandra>=3.7.0",
     "apache-airflow-providers-apache-drill>=2.8.1",
     "apache-airflow-providers-apache-druid>=3.12.0",
@@ -1378,7 +1374,6 @@ apache-airflow-shared-timezones = { workspace = true }
 apache-airflow-providers-airbyte = { workspace = true }
 apache-airflow-providers-alibaba = { workspace = true }
 apache-airflow-providers-amazon = { workspace = true }
-apache-airflow-providers-apache-beam = { workspace = true }
 apache-airflow-providers-apache-cassandra = { workspace = true }
 apache-airflow-providers-apache-drill = { workspace = true }
 apache-airflow-providers-apache-druid = { workspace = true }
@@ -1508,7 +1503,6 @@ members = [
     "providers/airbyte",
     "providers/alibaba",
     "providers/amazon",
-    "providers/apache/beam",
     "providers/apache/cassandra",
     "providers/apache/drill",
     "providers/apache/druid",
